Add PyTorch Lightning License
