+++ NOTICE
@@ -8,6 +8,11 @@ This project incorporates the following third-party components:
 - License: Apache License 2.0
 - License file: licenses/DINOv2_LICENSE
 
+**PyTorch Lightning**
+- Source: https://github.com/Lightning-AI/pytorch-lightning
+- License: Apache License 2.0
+- License file: licenses/PYTORCH_LIGHTNING_LICENSE
+
 ---
 
 All third-party code is used in accordance with their respective licenses.
