+++ src/lightly_train/_modules/teachers/dinov2/models/__init__.py
@@ -17,7 +17,7 @@ def build_model(
 ) -> Union[Tuple[Module, int], Tuple[Module, Module, int]]:
     suffix = "_memeff"
     if args.arch.endswith(suffix):
-        args.arch = args.arch[:-len(suffix)]
+        args.arch = args.arch[: -len(suffix)]
     if "vit" in args.arch:
         vit_kwargs = dict(
             img_size=img_size,
